+++ lib/feathericon/sass.rb
@@ -1,21 +1,8 @@
 require "feathericon/sass/engine"
+require 'feathericon/rails/railtie'
 
 module Feathericon
   module Sass
-    module Rails
-      module ViewHelpers
-        def icon(name, text = nil, html_options = {})
-          text, html_options = nil, text if text.is_a?(Hash)
-
-          content_class = "fe fa-#{name}"
-          content_class << " #{html_options[:class]}" if html_options.key?(:class)
-          html_options[:class] = content_class
-
-          html = content_tag(:i, nil, html_options)
-          html << ' ' << text.to_s unless text.blank?
-          html
-        end
-      end
-    end
+    # Your code goes here...
   end
 end
